+++ src/groups/power_group.jl
@@ -77,10 +77,10 @@ end
 @inline function active_traits(f, M::PowerGroup, args...)
     if is_metric_function(f)
         #pass to manifold by default - but keep Group Decorator for the retraction
-        return merge_traits(IsGroupManifold(M.op, M.gvr), IsExplicitDecorator())
+        return merge_traits(IsGroupManifold(M.op, M.vectors), IsExplicitDecorator())
     else
         return merge_traits(
-            IsGroupManifold(M.op, M.gvr),
+            IsGroupManifold(M.op, M.vectors),
             active_traits(f, M.manifold, args...),
             IsExplicitDecorator(),
         )